remove showcase customizations in templates
As we are now using the default routes created, we want to use the
default package controller templates, this also reduces and cleans up
the number of custom templates needed. As showcase had a lot of custom
templates for its relations etc. The default templates also make it
clearer what metadata fields we are adding in our schema